refactor: alterando os dtos de company e technology da pasta docs, para a pasta de dtos
