--- src/Docs/companys/index.ts
@@ -1,2 +1,0 @@
-export * from './create-company.dto';
-export * from './update-company.dto';
