+++ src/Docs/index.ts
@@ -1,3 +1,2 @@
 export * from './users';
-export * from './companys';
 export * from './auth';
