+++ src/companys/company.controller.ts
@@ -14,7 +14,7 @@ import {
 import { ApiTags } from '@nestjs/swagger';
 
 import { CompanyService } from './company.service';
-import { CreateCompanyDto } from '../Docs/companys/create-company.dto';
+import { CreateCompanyDto } from '../auth/Config/dtos/companys/create-company.dto';
 import { AuthGuard, RoleGuard, UserRoleEnum } from 'src/auth/Config';
 import { Roles } from 'src/auth/Config/decorators/roles.decorator';
 import { UpdateCompanyDto } from 'src/Docs';
